// fe-app-padkap-evolution/src/padkap-evolution/types.ts
export namespace PadkapEvolution {
// ...
  export interface MethodSuccessResponse<T> {
    success: true;
    data: T;
  }

  export interface MethodFailureResponse {
    success: false;
    error: string;
  }

  export type MethodResponse<T> =
    | MethodSuccessResponse<T>
    | MethodFailureResponse;
// ...
}